src/leafs_rice/requirements.txt
fastapi==0.87.0
python-multipart==0.0.5
uvicorn==0.19.0
pydantic
fastai==2.7.12
Pillow
timm==0.6.2.dev0